+++ CHANGELOG.md
@@ -22,6 +22,7 @@
 - Increase Slack task title size limit
 - Disable Todoist task search when not connected
 - Fix Slack message format with missing new lines
+- Consider API calls without change (304 status) as successful
 
 ## 2024-10-21